cli/requirements.txt
textual
httpx
rich
python-dotenv
